// packages-archived/coil-extension/src/types/commands.ts
export interface TipSent {
  command: 'tip'
  data: {
    paymentPointer: string
    amount: string
    assetCode: string
    assetScale: number
  }
}